Move logout to card
